Add log for sdw-config 0.5.0-rc1
Based on changes presented in

https://github.com/freedomofpress/securedrop-workstation/pull/619

Note that no tag is used here, since QA is happening against the feature
branch. In lieu of a tag, I verified a commit in the build logs, but
that commit hash will be lost with subsequent rebases. Still, the sig
verification is useful for review of the package.